// pronti-frontend-main/front/src/app/modulos/historial-laboral/form/form-historial-laboral.enum.ts
export enum FormHistorialLaboralEnum {
  // nombre = 'nombre',
  idCargo = 'idCargo',
  tipoContrato = 'tipoContrato',
  sueldo = 'sueldo',
  fechaIngreso = 'fechaIngreso',
  fechaFin = 'fechaFin',
  duracion = 'duracion',
  fueAscendido = 'fueAscendido',
  idAgencia = 'idAgencia',
  codigoSectorial = 'codigoSectorial',
  idDivisionAdministrativa = 'idDivisionAdministrativa'
}